+++ app/salary.py
@@ -1,0 +1,14 @@
+"""
+This module contains model SalaryIndex
+"""
+
+
+class SalaryIndex:
+    """
+    This class contains salary index and functions about this process
+    """
+    pass
+
+
+if __name__ == '__main__':
+    pass
